+++ src/college-sports/college-sports-home-page.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+import { HomePageTemplate } from "../home-page/tabs/home-page-template";
+import { OpinionsHeader } from "../home-page/tabs/common-components/opinions-header";
+
+export const CollegeSportsHomePage = () => {
+  return (
+    <HomePageTemplate headerLabel="College Sports">
+      <OpinionsHeader />
+    </HomePageTemplate>
+  );
+};
