--- src/home-page/tabs/tab-bodies/college-tab-body.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-import { BodyTemplate } from "./body-template";
-
-import { OpinionsHeader } from "../common-components/opinions-header";
-
-export const CollegeTabBody = () => {
-  return (
-    <BodyTemplate headerLabel="College Sports">
-      <OpinionsHeader />
-    </BodyTemplate>
-  );
-};
