--- src/home-page/tabs/tab-bodies/nfl-tab-body.tsx
@@ -1,11 +1,0 @@
-import React from "react";
-import { BodyTemplate } from "./body-template";
-import { OpinionsHeader } from "../common-components/opinions-header";
-
-export const NFLTabBody = () => {
-  return (
-    <BodyTemplate headerLabel="NFL">
-      <OpinionsHeader />
-    </BodyTemplate>
-  );
-};
